+++ netbox/netbox/settings.py
@@ -22,7 +22,7 @@ except ImportError:
     )
 
 
-VERSION = '2.5.5'
+VERSION = '2.5.6'
 
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))